# docs/api.yml
openapi: 3.0.3
info:
  title: API Server
  version: 0.0.4
servers:
  - url: https://boolean-api-server.fly.dev
...
paths:
  /groceries:
    get:
      tags:
        - groceries
      summary: Get a list of all available groceries
      description: Returns an array of groceries
      operationId: getGroceriesList
      parameters:
        - name: type
          in: query
          description: The type of items to return
          schema:
            type: string
            enum: [vegetable, fruit]
            example: vegetable
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grocery'
        '400':
          description: Failure - Invalid parameter provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /{username}/todo:
    post:
      tags:
        - todo
      summary: Add a new todo to the list
      description: Add a new todo to the list
      operationId: addTodo
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      requestBody:
        description: Create a new todo in the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
          application/xml:
            schema:
              $ref: '#/components/schemas/Todo'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Todo'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
            application/xml:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    get:
      tags:
        - todo
      summary: Get all of the todos in the list
      description: Returns an array of all of the existing todo items in the list
      operationId: getAllTodos
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    delete:
      tags:
        - todo
      summary: Delete all todo data
      description: Clear all todo data
      operationId: deleteAllTodos
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: No data to delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
...
  /{username}/pokemon:
    get:
      tags:
        - pokemon
      summary: Get all of the pokemon in the list
      description: Returns an array of all of the existing pokemon in the list
      operationId: getAllPokemon
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pokemon'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    post:
      tags:
        - pokemon
      summary: Add a new pokemon to the list
      description: Add a new pokemon to the list
      operationId: addPokemon
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      requestBody:
        description: Create a new pokemon in the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pokemon'
          application/xml:
            schema:
              $ref: '#/components/schemas/Pokemon'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Pokemon'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
            application/xml:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '400':
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    delete:
      tags:
        - pokemon
      summary: Delete all pokemon data
      description: Clear all pokemon data
      operationId: deleteAllPokemon
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: No data to delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
...
  /{username}/contact:
    post:
      tags:
        - contact
      summary: Add a new contact to the list
      description: Add a new contact to the list
      operationId: addContact
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      requestBody:
        description: Create a new contact in the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
          application/xml:
            schema:
              $ref: '#/components/schemas/Contact'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Contact'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
            application/xml:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    get:
      tags:
        - contact
      summary: Get all of the contacts in the list
      description: Returns an array of all of the existing contacts in the list
      operationId: getAllContacts
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    delete:
      tags:
        - contact
      summary: Delete all contact data
      description: Clear all contact data
      operationId: deleteAllContacts
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: No data to delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
...
  /{username}/post:
    post:
      tags:
        - post
      summary: Add a new post to the list
      description: Add a new post to the list
      operationId: addPost
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      requestBody:
        description: Create a new post in the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
          application/xml:
            schema:
              $ref: '#/components/schemas/Post'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Post'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
            application/xml:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    get:
      tags:
        - post
      summary: Get all of the posts in the list
      description: Returns an array of all of the existing posts in the list
      operationId: getAllPosts
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
    delete:
      tags:
        - post
      summary: Delete all post data
      description: Clear all post data
      operationId: deleteAllPosts
      parameters:
        - name: username
          in: path
          description: your github username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: No data to delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: The username provided is not part of the boolean-uk github organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubAuthIssue'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: invalid item type
    Message:
      type: object
      properties:
        message:
          type: string
          example: successful operation
    Grocery:
      type: object
      properties:
        id:
          type: string
          example: 006-bananas
        name:
          type: string
          example: bananas
        type:
          type: string
          enum: [vegetable, fruit]
          example: fruit
        price:
          type: number
          minimum: 0
          exclusiveMinimum: true
    Todo:
      required:
        - title
      type: object
      properties:
        title:
          type: string
          example: Clean my room
...
    Pokemon:
      required:
        - name
        - image
      type: object
      properties:
        name:
          type: string
          example: Pikachu
        image:
          type: string
          example: https://upload.wikimedia.org/wikipedia/en/thumb/a/a6/Pok%C3%A9mon_Pikachu_art.png/220px-Pok%C3%A9mon_Pikachu_art.png
        liked:
          type: boolean
          example: false
...
    Contact:
      required:
        - firstName
        - lastName
        - street
        - city
      type: object
      properties:
        firstName:
          type: string
          example: Rick
        lastName:
          type: string
          example: Sanchez
        street:
          type: string
          example: Morty Lane
        city:
          type: string
          example: Jerryville
        gender:
          type: string
          example: Male
        email:
          type: string
          example: rick@sanchez.com
        jobTitle:
          type: string
          example: Scientist
        latitude:
          type: number
          example: 42
        longitude:
          type: number
          example: 629
...
    Post:
      required:
        - title
        - content
        - contactId
      type: object
      properties:
        title:
          type: string
          example: I bought a new coffee machine
        content:
          type: string
          example: It's the best thing I've ever done!
        contactId:
          type: number
          example: 42
...
    BadRequest:
      type: object
      properties:
        error:
          type: array
          example:
            - body - title is required
            - params - id must be a number
          items:
            type: string
    GithubAuthIssue:
      type: object
      properties:
        error:
          type: string
          example: you must be part of the boolean-uk organisation to use this api
...